SW-5024-prep Refactor Votes pages for reusability
- Voting read-only and edit-views have similar data and scaffolding
- Extracted that out into a VotingWrapper
- Using a VotingProvider react context to provide the votes data as both views leverage url params
- removed the useFetchVotes hook
- TODO: implement voting edit details (Follow up)
